refactor: 피드 DTO 클래스명 리팩터링
