+++ src/main/java/com/souf/soufwebsite/domain/feed/controller/FeedController.java
@@ -9,10 +9,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.List;
 
 import static com.souf.soufwebsite.domain.feed.controller.FeedSuccessMessage.*;
 
@@ -24,11 +22,11 @@ public class FeedController {
     private final FeedService feedService;
 
     @PostMapping
-    public SuccessResponse<FeedCreateResDto> createFeed(
+    public SuccessResponse<FeedResDto> createFeed(
             @RequestBody @Valid FeedReqDto feedReqDto) {
-        FeedCreateResDto feedCreateResDto = feedService.createFeed(feedReqDto);
+        FeedResDto feedResDto = feedService.createFeed(feedReqDto);
 
-        return new SuccessResponse<>(feedCreateResDto, FEED_CREATE.getMessage());
+        return new SuccessResponse<>(feedResDto, FEED_CREATE.getMessage());
     }
 
     @GetMapping("/{memberId}")
@@ -39,7 +37,7 @@ public class FeedController {
     }
 
     @GetMapping("/{memberId}/{feedId}")
-    public SuccessResponse<FeedResDto> getFeed(
+    public SuccessResponse<FeedDetailResDto> getFeed(
             @PathVariable(name = "memberId") Long memberId,
             @PathVariable(name = "feedId") Long feedId) {
         return new SuccessResponse<>(feedService.getFeedById(memberId, feedId), FEED_GET.getMessage());
@@ -48,9 +46,8 @@ public class FeedController {
     @PatchMapping("/{feedId}")
     public SuccessResponse<?> updateFeed(
             @PathVariable Long feedId,
-            @RequestPart("data") @Valid FeedUpdateReqDto reqDto,
-            @RequestPart(name = "files", required = false) List<MultipartFile> newFiles) throws IOException {
-        feedService.updateFeed(feedId, reqDto, newFiles);
+            @RequestBody @Valid FeedReqDto reqDto) throws IOException {
+        feedService.updateFeed(feedId, reqDto);
         return new SuccessResponse<>(FEED_UPDATE.getMessage());
     }
