--- src/main/java/com/souf/soufwebsite/domain/feed/dto/FeedCreateResDto.java
@@ -1,12 +1,0 @@
-package com.souf.soufwebsite.domain.feed.dto;
-
-import com.souf.soufwebsite.domain.file.dto.PresignedUrlResDto;
-
-import java.util.List;
-
-public record FeedCreateResDto(
-        Long feedId,
-        List<PresignedUrlResDto> dtoList
-) {
-
-}
