+++ src/main/java/com/souf/soufwebsite/domain/feed/dto/FeedResDto.java
@@ -1,33 +1,12 @@
 package com.souf.soufwebsite.domain.feed.dto;
 
-import com.souf.soufwebsite.domain.feed.entity.Feed;
-import com.souf.soufwebsite.domain.file.dto.MediaResDto;
-import com.souf.soufwebsite.domain.file.entity.Media;
+import com.souf.soufwebsite.domain.file.dto.PresignedUrlResDto;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public record FeedResDto(
-        String topic,
-        String content,
-        int view,
-        List<MediaResDto> mediaResDtos,
-        LocalDateTime lastModifiedTime
+        Long feedId,
+        List<PresignedUrlResDto> dtoList
 ) {
-    public static FeedResDto from(Feed feed){
-        return new FeedResDto(
-                feed.getTopic(),
-            feed.getContent(),
-            feed.getViewCount(),
-            convertToMediaResDto(feed.getMedia()),
-            feed.getLastModifiedTime());
-    }
 
-    private static List<MediaResDto> convertToMediaResDto(List<Media> mediaList){
-        return mediaList.stream().map(
-                media -> MediaResDto.fromFeedDetail(media)
-        ).collect(Collectors.toList());
-    }
 }
