+++ src/main/java/com/souf/soufwebsite/domain/feed/service/FeedService.java
@@ -4,22 +4,20 @@ import com.souf.soufwebsite.domain.feed.dto.*;
 import com.souf.soufwebsite.domain.file.dto.MediaReqDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.List;
 
 public interface FeedService {
 
-    FeedCreateResDto createFeed(FeedReqDto reqDto);
+    FeedResDto createFeed(FeedReqDto reqDto);
 
     void uploadFeedMedia(MediaReqDto mediaReqDto);
 
     Page<FeedSimpleResDto> getFeeds(Long memberId, Pageable pageable);
 
-    FeedResDto getFeedById(Long memberId, Long feedId);
+    FeedDetailResDto getFeedById(Long memberId, Long feedId);
 
-    void updateFeed(Long feedId, FeedUpdateReqDto reqDto, List<MultipartFile> newFiles) throws IOException;
+    void updateFeed(Long feedId, FeedReqDto reqDto);
 
     void deleteFeed(Long feedId);
 }
